MapScript expanded and TurnManager works well
Made it so we can define the dimensions of the playing field from the turn manager. Therefore we could potentially allow the player to decide how big the map should be.

Alternatively we have a preset number of buttons with a 3x3 map, a 9x9 map, etc.

Camera needs to take into account how big the map is!