Basic framework for player input
Added sections in the TurnManager script that can be called on when the player tries to move and click.

the intended use is as follows:
Each turn the player gets to FIRST move a section of the labyrinth,
THEN they get to make a move of their own character to a valid tile (if possible).